+++ backend/src/main/java/org/timofeeva/docs/facade/impl/EmployeeFacadeImpl.java
@@ -7,6 +7,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Component;
 import org.timofeeva.docs.dto.EmployeeDTO;
 import org.timofeeva.docs.dto.EmployeeView;
+import org.timofeeva.docs.error.NotFoundException;
 import org.timofeeva.docs.facade.EmployeeFacade;
 import org.timofeeva.docs.mapper.EmployeeMapper;
 import org.timofeeva.docs.service.EmployeeService;
@@ -30,7 +31,9 @@ class EmployeeFacadeImpl implements EmployeeFacade {
 
     @Override
     public void deleteEmployee(Long id) {
-        service.delete(id);
+        if (service.existById(id)) {
+            service.delete(id);
+        } else throw new NotFoundException("id");
     }
 
 }
